test: split city credit Postgres suite by concern
